requirement.txt
gspread==5.12.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client
gspread_formatting
python-dateutil
